requirements.txt
hydra-core==1.3.2
imageio==2.31.1
matplotlib==3.7.1
networkx==2.8.7
numpy==1.23
omegaconf==2.3.0
overrides==7.3.1
pandas==1.4
pyemd==1.0.0
PyGSP==0.5.1
pytorch_lightning==2.0.4
scipy==1.11.0
setuptools==68.0.0
torch_geometric==2.3.1
torchmetrics==0.11.4
tqdm==4.65.0
wandb==0.15.4
gradio==4.20.0
pytdc
easydict